neuron/requirements.txt
-i https://pypi.org/simple
absl-py==0.9.0
apscheduler==3.6.0
astunparse==1.6.3
attrs==19.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
base58==2.0.1; python_version >= '3.5'
bitarray==1.2.2
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cytoolz==0.10.1; implementation_name == 'cpython'
eth-abi==2.1.1; python_version >= '3.6' and python_version < '4'
eth-account==0.5.2; python_version >= '3.6' and python_version < '4'
eth-hash[pycryptodome]==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1; python_version >= '3.5' and python_version < '4'
eth-utils==1.9.3; python_version >= '3.5' and python_full_version != '3.5.2' and python_version < '4'
flask-cors==3.0.8
flask==1.0.3
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.1
google-auth==1.20.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
hexbytes==0.2.1; python_version >= '3.6' and python_version < '4'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==1.7.0; python_version < '3.8' and python_version < '3.8'
ipfshttpclient==0.6.0.post1
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
jsonschema==3.2.0
keras-preprocessing==1.1.2
lru-dict==1.1.6
markdown==3.2.2; python_version >= '3.5'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
multiaddr==0.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
netaddr==0.8.0
numpy==1.16.4
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opt-einsum==3.3.0; python_version >= '3.5'
parsimonious==0.8.1
protobuf==3.12.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycryptodome==3.9.8
pyrsistent==0.16.0
python-dotenv==0.14.0
pytz==2020.1
requests-oauthlib==1.3.0
requests==2.24.0
rlp==1.2.0
rsa==4.6; python_version >= '3.5'
scipy==1.4.1; python_version >= '3.5'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
tensorboard-plugin-wit==1.7.0
tensorboard==2.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1'
tensorflow-estimator==2.3.0
tensorflow==2.3.0
termcolor==1.1.0
toolz==0.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.7.4.2; python_version < '3.8'
tzlocal==2.1
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
varint==1.0.2
web3==5.12.0
websockets==8.1; python_full_version >= '3.6.1'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.34.2; python_version >= '3'
wrapt==1.12.1
zipp==3.1.0; python_version >= '3.6'
